// src/app/model/ListUser.ts
export interface ListUser {
  appliedInstance: string;
  username: string;
  userID: string;
  user_Profile?: string;
  registering?: string;
  fingerprint: string;
  email: string;
  expired_Password: string;
  approval_Notes: string;
  ipAddress: string;
  password: string;
  retypePassword: string;
  status: string;
  lastChange: string;
  description: string;
}